requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
atlastk==0.13.2
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.2
cryptography==37.0.4
DateTime==4.5
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-pagination==0.9.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
kombu==5.2.4
lxml==4.9.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
prompt-toolkit==3.0.30
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.1
tortoise==0.1.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
vine==5.0.0
watchfiles==0.15.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
zope.interface==5.4.0
